requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-cleanup==6.0.0
djangorestframework==3.12.4
drf-yasg==1.21.4
environs==9.5.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
Pillow==8.4.0
psycopg2==2.9.4
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.8
